+++ ci-scripts/test/integration/pingaccess/05-test-cloudwatch-logs.sh
@@ -8,8 +8,8 @@ if skipTest "${0}"; then
 fi
 
 oneTimeSetUp() {
-  readonly PA_ADMIN_LOG_STREAM_SUFFIX="pingaccess-admin-0_${NAMESPACE}_pingaccess-admin"
-  readonly PA_ENGINE_LOG_STREAM_SUFFIX="pingaccess-0_${NAMESPACE}_pingaccess"
+  readonly PA_ADMIN_LOG_STREAM_SUFFIX="pingaccess-admin-0_${PING_CLOUD_NAMESPACE}_pingaccess-admin"
+  readonly PA_ENGINE_LOG_STREAM_SUFFIX="pingaccess-0_${PING_CLOUD_NAMESPACE}_pingaccess"
 }
 
 testPaLogStreamsExist() {