+++ ci-scripts/test/integration/pingcentral/prerequisites/01-verify-pods.sh
@@ -15,10 +15,10 @@ testPodConnection() {
   # Get pingcentral pod name
   pingcentral_pod_name=$(kubectl get pods \
                           -l "${pod_label_name}" \
-                          -n "${NAMESPACE}" \
+                          -n "${PING_CLOUD_NAMESPACE}" \
                           -o=jsonpath="{.items[*].metadata.name}" | tr -s '[[:space:]]')
   pingcentral_ready_state=$(kubectl get pods "${pingcentral_pod_name}" \
-                            -n "${NAMESPACE}" | tail -n +2 | awk '{print $2}' | tr -s '[[:space:]]')
+                            -n "${PING_CLOUD_NAMESPACE}" | tail -n +2 | awk '{print $2}' | tr -s '[[:space:]]')
   assertEquals "Failed to get pingcentral running state 1/1" "${expected_ready_state}" "${pingcentral_ready_state}"
 }